+++ packages/f-metadata/CHANGELOG.md
@@ -3,6 +3,18 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+2.2.0
+------------------------------
+*February  20, 2020*
+
+### Added
+- Jest configuration and unit tests for both server and client.
+
+### Changed
+- Promisify `f-metadata` initialisation and improve error handling.
+- Callback resolve with the appboy instance.
+
+
 2.1.0
 ------------------------------
 *February  18, 2020*